requirements.txt
numpy
pandas
jupyter
psycopg2
tqdm
pyyaml
openpyxl